chore: convert @packages/example to TypeScript
Convert the example package's source to TypeScript:

- `lib/example.js` + `lib/example.d.ts` collapse into a single
  `lib/example.ts`. It uses `export =` so `require('@packages/example')`
  keeps returning `{ getPathToE2E }` exactly as before; `index.d.ts`
  now re-exports that assignment instead of a synthetic default.
- `bin/build.js` and `bin/convert.js` become `bin/build.ts` and
  `bin/convert.ts`, run with `tsx` to match how the other packages run
  their build scripts. `build` now resolves `cypress-example-kitchensink`
  up front and fails with an actionable message when it is missing,
  rather than passing `undefined` into `path.join` after it has already
  removed `app/` and `cypress/`.
- Add a `tsconfig.json` and a `check-ts` script so the package is
  type-checked, and declare the `fs-extra`/`glob` dependencies the build
  scripts were previously picking up from the workspace root.

`gulpfile.js` and `cypress.config.js` stay JavaScript — gulp and Cypress
load them directly without a TypeScript hook.

Verified `yarn workspace @packages/example build` produces byte-identical
`app/` and `cypress/` output to the previous scripts.
