+++ packages/example/index.d.ts
@@ -1,1 +1,3 @@
-export { default } from './lib/example'
+import example = require('./lib/example')
+
+export = example
