--- packages/example/lib/example.d.ts
@@ -1,5 +1,0 @@
-declare const example: {
-  getPathToE2E(): string
-}
-
-export default example
