--- packages/example/lib/example.js
@@ -1,7 +1,0 @@
-const path = require('path')
-
-module.exports = {
-  getPathToE2E () {
-    return path.join(__dirname, '..', 'cypress', 'e2e')
-  },
-}
