+++ packages/example/package.json
@@ -4,7 +4,8 @@
   "private": true,
   "main": "index.js",
   "scripts": {
-    "build": "node ./bin/build.js && gulp build",
+    "build": "tsx ./bin/build.ts && gulp build",
+    "check-ts": "tsc --noEmit",
     "clean-deps": "rimraf node_modules",
     "predeploy": "yarn build",
     "deploy": "gh-pages -d build -b gh-pages",
@@ -14,12 +15,17 @@
     "test-unit": "echo 'no unit tests'"
   },
   "devDependencies": {
+    "@types/fs-extra": "^9.0.13",
+    "@types/glob": "7.1.1",
     "cypress-example-kitchensink": "6.1.0",
+    "fs-extra": "9.1.0",
     "gh-pages": "5.0.0",
+    "glob": "7.1.6",
     "gulp": "4.0.2",
     "gulp-clean": "0.4.0",
     "gulp-rev-all": "2.0.2",
-    "resolve-pkg": "2.0.0"
+    "resolve-pkg": "2.0.0",
+    "tsx": "^4.23.12"
   },
   "files": [
     "cypress",
